Fix several integer conversions in process_ops
Fix UsersTest.test_sanity on Windows.

uid and gid were returned as int (while they normally are unsigned int)
and converted to signed integers in the table row.
This is wrong because beyond uid and gid not being ints,
they are taken from the RID part of the SID which in some cases,
like for a Service SID, it can have a value higher than then maximum
value of an int, so in the end the number shown in table is negative.

Now they are returned as uint32_t and converted as BIGINTs for the table
that uses them.

Fix other functions return values and conversions depending on the meaning of
the value.
On Windows stick to its specific types where possible.

Convert CRLF to LF on some of the files modified.

+++ osquery/process/windows/process_ops.h
@@ -41,6 +41,6 @@ std::unique_ptr<BYTE[]> getSidFromUsername(std::wstring accountName);
  *
  * @returns the RID represented as an unsigned long integer.
  */
-unsigned long getRidFromSid(PSID sidPtr);
+DWORD getRidFromSid(PSID sidPtr);
 
 } // namespace osquery
